requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
email-validator==2.1.0
websockets==12.0
python-socketio==5.11.0
scikit-learn==1.4.0
numpy==1.26.3
bcrypt==4.1.2
xlsxwriter==3.1.9
reportlab==4.0.8
psycopg2-binary==2.9.9
python-dotenv==1.0.1